requirements.txt
cx_Oracle
psycopg2-binary
pytest
